+++ devel/load_gillespie_example_figure.py
@@ -6,7 +6,7 @@ import base64
 
 import kachery as ka
 import figpack.views as fpv
-import figpack.franklab.views as flv
+from figpack_franklab import TrackAnimation
 
 
 def decode_base64_to_numpy(data_b64, dtype, shape):
@@ -129,7 +129,7 @@ def process_track_animation(data):
     print(f"ymax: {metadata['ymax']}")
     print(f"ymin: {metadata['ymin']}")
 
-    X = flv.TrackAnimation(
+    X = TrackAnimation(
         bin_height=decoded_pos["bin_height"],
         bin_width=decoded_pos["bin_width"],
         frame_bounds=decoded_pos["frame_bounds"],