+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add unit locations spike sorting view
 - Rework context provider structure in preparation for moving plugins to figpack extensions
+- Move franklab view to franklab_extension package
+- Remove extension registry (no longer needed)
 
 ## [0.2.15] - 2025-09-08
 - Support for Zarr v3 in addition to v2.